% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R.R
\name{import_data}
\alias{import_data}
\title{Charger et fusionner les données}
\usage{
import_data(shp_path, csv_path)
}
\arguments{
\item{shp_path}{Chemin vers le shapefile (.shp)}

\item{csv_path}{Chemin vers le CSV contenant la consommation}
}
\value{
Un objet sf fusionné
}
\description{
Charger et fusionner les données
}
