% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R.R
\name{plot_map}
\alias{plot_map}
\title{Afficher une carte des quantités consommées}
\usage{
plot_map(merged_data)
}
\arguments{
\item{merged_data}{Un objet sf contenant les régions et les quantités}
}
\description{
Afficher une carte des quantités consommées
}
